% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAD.R
\name{CAD}
\alias{CAD}
\title{CAD: Solar System Dynamics and Center for Near-Earth Object Studies - Asteroid and comet close approaches to the planets in the past and future}
\usage{
CAD(date_min = "now", date_max = lubridate::today() + 60,
  dist_min = NULL, dist_max = "0.05", h_min = NULL, h_max = NULL,
  v_inf_min = NULL, v_inf_max = NULL, v_rel_min = NULL,
  v_rel_max = NULL, class = NULL, pha = FALSE, nea = FALSE,
  comet = FALSE, nea_comet = FALSE, neo = TRUE, kind = NULL,
  spk = NULL, des = NULL, body = "Earth", sort = "date",
  limit = NULL, fullname = FALSE)
}
\arguments{
\item{date_min}{Date. Exclude data earlier than this date "YYYY-MM-DD" or date/time "YYYY-MM-DDThh:mm:ss" or "now" for the current date. "now" as default.}

\item{date_max}{Date. Exclude data later than this date "YYYY-MM-DD" or date/time "YYYY-MM-DDThh:mm:ss". "now" + 60 as default.}

\item{dist_max}{String. Exclude data with an approach distance greater than this (see dist-min). "0.05" as default.}

\item{h_min}{Number. Exclude data from objects with H-values less than this (e.g., 22 meaning objects smaller than this).}

\item{h_max}{Number. Exclude data from objects with H-value greater than this (e.g., 17.75 meaning objects larger than this).}

\item{v_inf_min}{Number. Exclude data with V-infinity less than this positive value in km/s (e.g., 18.5).}

\item{v_inf_max}{Number. Exclude data with V-infinity greater than this positive value in km/s (e.g., 20).}

\item{v_rel_min}{Number. Exclude data with V-relative less than this positive value in km/s (e.g., 11.2).}

\item{v_rel_max}{Number. Exclude data with V-relative greater than this positive value in km/s (e.g., 19).}

\item{class}{String. Limit data to objects with the specified orbit-class (e.g., "ATE").}

\item{pha}{Boolean. Limit data to PHAs. FALSE as default.}

\item{nea}{Boolean. Limit data to NEAs. FALSE as default.}

\item{comet}{Boolean. Limit data to comets. FALSE as default.}

\item{neo}{Boolean. Limit data to NEOs. TRUE as default.}

\item{kind}{String. Limit data to objects of the specified kind ("a"=asteriod, "an"=numbered-asteroids, "au"=unnumbered-asteroids, "c"=comets, "cn"=numbered-comets, "cu"=unnumbered-comets, "n"=numbered-objects, and "u"=unnumbered-objects).}

\item{spk}{Integer. Only data for the object matching this SPK-ID (e.g., 2000433).}

\item{des}{String. only data for the object matching this designation (e.g., "2015 AB" or "141P" or "433").}

\item{body}{String. Limit data to close-approaches to the specified body (e.g., "Earth") or allow all bodies with "ALL" or *. "Earth" as default.}

\item{sort}{String. Sort data on the specified field: “date”, “dist”, “dist-min”, “v-inf”, “v-rel”, “h”, or “object” (default sort order is ascending: prepend “-“ for descending). "date" as default.}

\item{limit}{Integer. Limit data to the first N results (where N is the specified number and must be an integer value greater than zero).}

\item{fullname}{Boolean. Include the full-format object name/designation. FALSE as default.}

\item{energy_min}{String. Exclude data with an approach distance less than this, e.g., "0.05", "10LD".}

\item{nea-comet}{Boolean. Limit data to NEAs and comets. FALSE as default.}
}
\value{
Data of asteroids and comets in JPL’s SBDB (Small-Body DataBase).
}
\description{
This API provides access to current close-approach data for all
asteroids and comets in JPL’s SBDB (Small-Body DataBase). See detailed info at: https://ssd-api.jpl.nasa.gov/doc/cad.html.
}
\examples{
CAD(des = "433", date_min = as.Date("2019-01-01"), date_max = as.Date("2100-01-01"), dist_max = "0.2")
CAD(dist_max = "10LD", date_min = as.Date("2018-01-01"), sort = "dist")
}
