% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DONKI_CME.R
\name{DONKI_CME}
\alias{DONKI_CME}
\title{DONKI_CME: Space Weather Database Of Notifications, Knowledge, Information - Coronal Mass Ejection (CME)}
\usage{
DONKI_CME(key = Sys.getenv("NASA_TOKEN"), start_date = end_date - 30,
  end_date = lubridate::today(tzone = "UTC"))
}
\arguments{
\item{key}{String. Your NASA API key, you can enter your key in the function parameter, but it's not recommended. Instead, you'd better save your key in R environment and call it "NASA_TOKEN". Then the function would automatically acquire your key info.}

\item{start_date}{Date. Starting UTC date for CME search. 30 days prior to current UTC date as default.}

\item{end_date}{Date. Ending UTC date for CME search. Current UTC date as default.}
}
\value{
Data of CME.
}
\description{
Get access to the data of Coronal Mass Ejection (CME).
}
\examples{
DONKI_CME(end_date = as.Date("2019-11-20"))
}
