% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DONKI_CME_Analysis.R
\name{DONKI_CME_Analysis}
\alias{DONKI_CME_Analysis}
\title{DONKI_CME: Space Weather Database Of Notifications, Knowledge, Information - Coronal Mass Ejection (CME) Analysis}
\usage{
DONKI_CME_Analysis(key = Sys.getenv("NASA_TOKEN"),
  start_date = end_date - 30, end_date = lubridate::today(tzone =
  "UTC"), mostAccurateOnly = TRUE, completeEntryOnly = TRUE,
  speed = 0, halfAngle = 0, catalog = "ALL", keyword = "NONE")
}
\arguments{
\item{key}{String. Your NASA API key, you can enter your key in the function parameter, but it's not recommended. Instead, you'd better save your key in R environment and call it "NASA_TOKEN". Then the function would automatically acquire your key info.}

\item{start_date}{Date. Starting UTC date for CME search. 30 days prior to current UTC date as default.}

\item{end_date}{Date. Ending UTC date for CME search. Current UTC date as default.}

\item{mostAccurateOnly}{Boolean. TRUE as default.}

\item{completeEntryOnly}{Boolean. TRUE as default.}

\item{speed}{Integer. 0 as default.}

\item{halfAngle}{Integer. 0 as default.}

\item{catalog}{String. "ALL" as default (choices: "ALL", "SWRC_CATALOG", "JANG_ET_AL_CATALOG").}

\item{keyword}{String. "NONE" as default (example choices: "swpc_annex").}
}
\value{
Data of CME Analysis.
}
\description{
Get access to the data of Coronal Mass Ejection (CME) Analysis.
}
\examples{
DONKI_CME_Analysis(start_date = as.Date("2016-09-01"), end_date = as.Date("2016-09-30"), speed = 500, halfAngle = 30)
}
