% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DONKI_FLR.R
\name{DONKI_FLR}
\alias{DONKI_FLR}
\title{DONKI_FLR: Space Weather Database Of Notifications, Knowledge, Information - Solar Flare (FLR)}
\usage{
DONKI_FLR(key = Sys.getenv("NASA_TOKEN"), start_date = end_date - 30,
  end_date = lubridate::today(tzone = "UTC"))
}
\arguments{
\item{key}{String. Your NASA API key, you can enter your key in the function parameter, but it's not recommended. Instead, you'd better save your key in R environment and call it "NASA_TOKEN". Then the function would automatically acquire your key info.}

\item{start_date}{Date. Starting UTC date for FLR search. 30 days prior to current UTC date as default.}

\item{end_date}{Date. Ending UTC date for FLR search. Current UTC date as default.}
}
\value{
Data of FLR.
}
\description{
Get access to the data of Solar Flare (FLR).
}
\examples{
DONKI_FLR(start_date = as.Date("2019-01-01"), end_date = as.Date("2019-06-30"))
}
