% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DONKI_GST.R
\name{DONKI_GST}
\alias{DONKI_GST}
\title{DONKI_GST: Space Weather Database Of Notifications, Knowledge, Information - Geomagnetic Storm (GST)}
\usage{
DONKI_GST(key = Sys.getenv("NASA_TOKEN"), start_date = end_date - 30,
  end_date = lubridate::today(tzone = "UTC"))
}
\arguments{
\item{key}{String. Your NASA API key, you can enter your key in the function parameter, but it's not recommended. Instead, you'd better save your key in R environment and call it "NASA_TOKEN". Then the function would automatically acquire your key info.}

\item{start_date}{Date. Starting UTC date for GST search. 30 days prior to current UTC date as default.}

\item{end_date}{Date. Ending UTC date for GST search. Current UTC date as default.}
}
\value{
Data of GST.
}
\description{
Get access to the data of Geomagnetic Storm (GST).
}
\examples{
DONKI_GST(start_date = as.Date("2015-01-01"), end_date = as.Date("2015-06-30"))
}
