% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DONKI_HSS.R
\name{DONKI_HSS}
\alias{DONKI_HSS}
\title{DONKI_HSS: Space Weather Database Of Notifications, Knowledge, Information - Hight Speed Stream (HSS)}
\usage{
DONKI_HSS(key = Sys.getenv("NASA_TOKEN"), start_date = end_date - 30,
  end_date = lubridate::today(tzone = "UTC"))
}
\arguments{
\item{key}{String. Your NASA API key, you can enter your key in the function parameter, but it's not recommended. Instead, you'd better save your key in R environment and call it "NASA_TOKEN". Then the function would automatically acquire your key info.}

\item{start_date}{Date. Starting UTC date for HSS search. 30 days prior to current UTC date as default.}

\item{end_date}{Date. Ending UTC date for HSS search. Current UTC date as default.}
}
\value{
Data of HSS.
}
\description{
Get access to the data of Hight Speed Stream (HSS).
}
\examples{
DONKI_HSS(start_date = as.Date("2019-01-01"))
}
