% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DONKI_IPS.R
\name{DONKI_IPS}
\alias{DONKI_IPS}
\title{DONKI_IPS: Space Weather Database Of Notifications, Knowledge, Information - Interplanetary Shock (IPS)}
\usage{
DONKI_IPS(key = Sys.getenv("NASA_TOKEN"), start_date = end_date - 30,
  end_date = lubridate::today(tzone = "UTC"), location = "ALL",
  catalog = "ALL")
}
\arguments{
\item{key}{String. Your NASA API key, you can enter your key in the function parameter, but it's not recommended. Instead, you'd better save your key in R environment and call it "NASA_TOKEN". Then the function would automatically acquire your key info.}

\item{start_date}{Date. Starting UTC date for IPS search. 30 days prior to current UTC date as default.}

\item{end_date}{Date. Ending UTC date for IPS search. Current UTC date as default.}

\item{location}{String. "ALL" as default (choices: "Earth", "MESSENGER", "STEREO A", "STEREO B").}

\item{catalog}{String. "ALL" as default (choices: "SWRC_CATALOG", "WINSLOW_MESSENGER_ICME_CATALOG").}
}
\value{
Data of IPS.
}
\description{
Get access to the data of Interplanetary Shock (IPS).
}
\examples{
DONKI_IPS(start_date = as.Date("2019-01-01"), end_date = as.Date("2019-10-31"), location = "Earth")
}
