% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DONKI_MPC.R
\name{DONKI_MPC}
\alias{DONKI_MPC}
\title{DONKI_MPC: Space Weather Database Of Notifications, Knowledge, Information - Magnetopause Crossing (MPC)}
\usage{
DONKI_MPC(key = Sys.getenv("NASA_TOKEN"), start_date = end_date - 30,
  end_date = lubridate::today(tzone = "UTC"))
}
\arguments{
\item{key}{String. Your NASA API key, you can enter your key in the function parameter, but it's not recommended. Instead, you'd better save your key in R environment and call it "NASA_TOKEN". Then the function would automatically acquire your key info.}

\item{start_date}{Date. Starting UTC date for MPC search. 30 days prior to current UTC date as default.}

\item{end_date}{Date. Ending UTC date for MPC search. Current UTC date as default.}
}
\value{
Data of MPC.
}
\description{
Get access to the data of Magnetopause Crossing (MPC).
}
\examples{
DONKI_MPC(start_date = as.Date("2018-01-01"))
}
