% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DONKI_Notifications.R
\name{DONKI_Notifications}
\alias{DONKI_Notifications}
\title{DONKI_Notifications: Space Weather Database Of Notifications, Knowledge, Information - Notifications}
\usage{
DONKI_Notifications(key = Sys.getenv("NASA_TOKEN"),
  start_date = end_date - 7, end_date = lubridate::today(tzone =
  "UTC"), type = "all")
}
\arguments{
\item{key}{String. Your NASA API key, you can enter your key in the function parameter, but it's not recommended. Instead, you'd better save your key in R environment and call it "NASA_TOKEN". Then the function would automatically acquire your key info.}

\item{start_date}{Date. Starting UTC date for Notifications search. 7 days prior to current UTC date as default. The request date range is limit to 30 days. If the request range is greater than 30 days, it would limit your request range to (end_date-30) to end_date.}

\item{end_date}{Date. Ending UTC date for Notifications search. Current UTC date as default.}

\item{type}{String. "all" as default (choices: "all", "FLR", "SEP", "CME", "IPS", "MPC", "GST", "RBE", "report")}
}
\value{
Data of Notifications.
}
\description{
Get access to the data of Notifications.
}
\examples{
DONKI_Notifications(start_date = as.Date("2019-01-01"), end_date = as.Date("2019-03-01"))
}
