% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DONKI_SEP.R
\name{DONKI_SEP}
\alias{DONKI_SEP}
\title{DONKI_SEP: Space Weather Database Of Notifications, Knowledge, Information - Solar Energetic Particle (SEP)}
\usage{
DONKI_SEP(key = Sys.getenv("NASA_TOKEN"), start_date = end_date - 30,
  end_date = lubridate::today(tzone = "UTC"))
}
\arguments{
\item{key}{String. Your NASA API key, you can enter your key in the function parameter, but it's not recommended. Instead, you'd better save your key in R environment and call it "NASA_TOKEN". Then the function would automatically acquire your key info.}

\item{start_date}{Date. Starting UTC date for SEP search. 30 days prior to current UTC date as default.}

\item{end_date}{Date. Ending UTC date for SEP search. Current UTC date as default.}
}
\value{
Data of SEP.
}
\description{
Get access to the data of Solar Energetic Particle (SEP).
}
\examples{
DONKI_SEP(start_date = as.Date("2000-01-01"), end_date = as.Date("2010-12-31"))
}
