% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DONKI_WSAEnlilSimulation.R
\name{DONKI_WSAEnlilSimulation}
\alias{DONKI_WSAEnlilSimulation}
\title{DONKI_WSAEnlilSimulation: Space Weather Database Of Notifications, Knowledge, Information - WSA+EnlilSimulation}
\usage{
DONKI_WSAEnlilSimulation(key = Sys.getenv("NASA_TOKEN"),
  start_date = end_date - 7, end_date = lubridate::today(tzone =
  "UTC"))
}
\arguments{
\item{key}{String. Your NASA API key, you can enter your key in the function parameter, but it's not recommended. Instead, you'd better save your key in R environment and call it "NASA_TOKEN". Then the function would automatically acquire your key info.}

\item{start_date}{Date. Starting UTC date for WSAEnlilSimulation search. 7 days prior to current UTC date as default.}

\item{end_date}{Date. Ending UTC date for WSAEnlilSimulation search. Current UTC date as default.}
}
\value{
Data of WSAEnlilSimulation.
}
\description{
Get access to the data of WSA+EnlilSimulation.
}
\examples{
DONKI_WSAEnlilSimulation(start_date = as.Date("2019-10-01"), end_date = as.Date("2019-10-15"))
}
