% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earth_Assets.R
\name{Earth_Assets}
\alias{Earth_Assets}
\title{Earth_Assets: The date-times and asset names for available imagery for a supplied location}
\usage{
Earth_Assets(key = Sys.getenv("NASA_TOKEN"), lat, lon, begin_date,
  end_date = Sys.Date())
}
\arguments{
\item{key}{String. Your NASA API key, you can enter your key in the function parameter, but it's not recommended. Instead, you'd better save your key in R environment and call it "NASA_TOKEN". Then the function would automatically acquire your key info.}

\item{lat}{Float. Latitude of the location.}

\item{lon}{Float. Longitude of the location.}

\item{begin_date}{Date. Beginning of date range.}

\item{end_date}{Date. End of date range. Today as default.}
}
\value{
A dataframe including date-times and asset names.
}
\description{
This function retrieves retrieves the date-times and asset names for available imagery for a supplied location.
The satellite passes over each point on earth roughly once every sixteen days.
This is an amazing visualization of the acquisition pattern for Landsat 8 imagery.
}
\examples{
Earth_Assets(lon = 100.75, lat = 1.5, begin_date = as.Date("2014-02-01"))
}
