% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earth_Imagery.R
\name{Earth_Imagery}
\alias{Earth_Imagery}
\title{Earth_Imagery: The Landsat 8 image on the Earth for the supplied location and date}
\usage{
Earth_Imagery(key = Sys.getenv("NASA_TOKEN"), lat, lon, dim = 0.025,
  date = Sys.Date(), cloud_score = FALSE,
  download_filename = "img.png")
}
\arguments{
\item{key}{String. Your NASA API key, you can enter your key in the function parameter, but it's not recommended. Instead, you'd better save your key in R environment and call it "NASA_TOKEN". Then the function would automatically acquire your key info.}

\item{lat}{Float. Latitude of the location.}

\item{lon}{Float. Longitude of the location.}

\item{dim}{Float. Width and height of image in degrees. 0.025 as default.}

\item{date}{Date. Date of image; if not supplied, then the most recent image (i.e., closest to today) is returned. Today as default.}

\item{cloud_score}{Boolean. Calculate the percentage of the image covered by clouds. FALSE as default. If False is supplied, then no keypair is returned. If True is supplied, then a keypair will always be returned, even if the backend algorithm is not able to calculate a score. Note that this is a rough calculation, mainly used to filter out exceedingly cloudy images.}

\item{download_filename}{String. Filename of downloaded image. "img.png" as default.}
}
\value{
A Landsat 8 imagery and its date.
}
\description{
This function retrieves the Landsat 8 image for the supplied location and date.
The response will include the date and URL to the image that is closest to the supplied date.
The requested resource may not be available for the exact date in the request.
}
\examples{
Earth_Imagery(lon = 100.75, lat = 1.5, date = as.Date("2014-02-01"), cloud_score = TRUE)
}
