% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EONET_Events.R
\name{EONET_Events}
\alias{EONET_Events}
\title{EONET_Events: The Earth Observatory Natural Event Tracker}
\usage{
EONET_Events(source = c("InciWeb", "EO"), status = "open", limit = 5,
  days = 20, by_events = TRUE, by_layers = FALSE, category_id = 8)
}
\arguments{
\item{source}{Vector. Event sources in the EONET system. c("InciWeb", "EO") as default. See detailed info: https://eonet.sci.gsfc.nasa.gov/api/v2.1/sources}

\item{status}{String. Events that have ended are assigned a closed date and the existence of that date will allow you to filter for only-open or only-closed events. "open" as default.}

\item{limit}{Integer. Limits the number of events returned. 5 as default.}

\item{days}{Integer. Limits the number of prior days (including today) from which events will be returned. 20 as default.}

\item{by_events}{Boolean. Filters the data by events(TRUE) or by categories(FALSE). TRUE as default.}

\item{by_layers}{Boolean. Filters the data by layers. FALSE as default.}

\item{category_id}{Integer. Filters the events by the category. 8 as default. See detailed info: https://eonet.sci.gsfc.nasa.gov/api/v2.1/categories}
}
\value{
List of info about natural events on the Earth.
}
\description{
Get access to the information of events on the Earth (filtered by events/categories/layers).
}
\details{
If you want to filter the data by events, then you need set these parameters: source, status, limit, days, by_events=TRUE(default).

If you want to filter the data by categories, then you need set these parameters: category_id, source, status, limit, days, by_events=FALSE.

If you want to filter the data by layers, then you need to set the parameter: by_layers = TRUE, category_id
}
\examples{
#### Filtered by events:
EONET_Events(source = "InciWeb", limit = 10, days = 30)
#### Filtered by categories:
EONET_Events(category_id = 8, source = "EO", limit = 5, days = 365, by_events = FALSE)
#### Filtered by layers:
EONET_Events(by_layers = TRUE, category_id = 14)
}
