% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exoplanet.R
\name{Exoplanet}
\alias{Exoplanet}
\title{Exoplanet: Access to NASA's Exoplanet Archive database}
\usage{
Exoplanet(table = "exoplanets", select = NULL, count = NULL,
  colset = NULL, where = NULL, order = NULL, ra = NULL,
  dec = NULL, radius = NULL, aliastable = NULL)
}
\arguments{
\item{table}{String. Specifies which table to query. "exoplanets" as default. Check the website for a full list of tables.}

\item{select}{String. Specifies which columns within the chosen table to return. Check the website for valid column names.}

\item{count}{Integer. Can be used to return the number of rows which fulfill the given query, including queries using where clauses or cone searches.}

\item{colset}{String. Returns a set of pre-defined columns that have been created by the archive. Currently, this keyword is only used by the Composite Planet Data (compositepars) table.}

\item{where}{String/A vector of strings. Specifies which rows to return. Use this to search for a range of values, such as rows with a declination greater than 0. Parameters must use a valid column name.}

\item{order}{String. Controls the order the rows are returned.}

\item{ra}{Number. Specifies an area of the sky to search for all objects within that area. Right Ascension (ra), Declination (dec) and radius (radius or rad) must be listed with their respective coordinates or units.}

\item{dec}{Number. Specifies an area of the sky to search for all objects within that area. Right Ascension (ra), Declination (dec) and radius (radius or rad) must be listed with their respective coordinates or units.}

\item{radius}{String. Specifies an area of the sky to search for all objects within that area. Right Ascension (ra), Declination (dec) and radius (radius or rad) must be listed with their respective coordinates or units. The units must be included in the radius specification (degrees, minutes, arcsecs) with a space preceding the unit name.}

\item{aliastable}{String. Requests a list of aliases for a particular confirmed planet.}
}
\value{
Data from NASA's Exoplanet Archive database.
}
\description{
This function allows programatic access to NASA's Exoplanet Archive database.
See detailed info at: https://exoplanetarchive.ipac.caltech.edu/docs/program_interfaces.html#data.
}
\examples{
Exoplanet()
Exoplanet(select = "pl_hostname", order = "dec")
Exoplanet(ra = 291, dec = 48, radius = "1 degree")
Exoplanet(table = "cumulative", where = c("koi_prad<2", "koi_teq>180", "koi_teq<303", "koi_disposition like 'CANDIDATE'"))
Exoplanet(aliastable = "bet Pic")
}
