% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fireball.R
\name{Fireball}
\alias{Fireball}
\title{Fireball: Solar System Dynamics and Center for Near-Earth Object Studies - Fireball atmospheric impact data reported by US Government sensors}
\usage{
Fireball(date_min = lubridate::today(), date_max = lubridate::today() +
  60, energy_min = NULL, energy_max = NULL, impact_e_min = NULL,
  impact_e_max = NULL, vel_min = NULL, vel_max = NULL,
  alt_min = NULL, alt_max = NULL, req_loc = FALSE, req_alt = FALSE,
  req_vel = FALSE, req_vel_comp = FALSE, vel_comp = FALSE,
  sort = "date", limit = NULL)
}
\arguments{
\item{date_min}{Date. Exclude data earlier than this date "YYYY-MM-DD" or date/time "YYYY-MM-DDThh:mm:ss". Today as default.}

\item{date_max}{Date. Exclude data later than this date "YYYY-MM-DD" or date/time "YYYY-MM-DDThh:mm:ss". "now" + 60 as default.}

\item{energy_min}{String. Exclude data with total-radiated-energy less than this positive value in joules ×10^10 (e.g., "0.3" = 0.3×1010 joules).}

\item{energy_max}{String. Exclude data with total-radiated-energy greater than this (see energy-min).}

\item{impact_e_min}{String. Exclude data with estimated impact energy less than this positive value in kilotons (kt) (e.g., "0.08" kt).}

\item{impact_e_max}{String. Exclude data with total-radiated-energy greater than this (see impact-e-min).}

\item{vel_min}{Number. Exclude data with velocity-at-peak-brightness less than this positive value in km/s (e.g., 18.5).}

\item{vel_max}{Number. Exclude data with velocity-at-peak-brightness greater than this positive value in km/s (e.g., 20).}

\item{alt_min}{Number. Exclude data from objects with an altitude less than this (e.g., 22 meaning objects smaller than this).}

\item{alt_max}{Number. Exclude data from objects with an altitude greater than this (e.g., 17.75 meaning objects larger than this).}

\item{req_loc}{Boolean. Location (latitude and longitude) required; when set TRUE, exclude data without a location. FALSE as default.}

\item{req_alt}{Boolean. Altitude required; when set true, exclude data without an altitude. FALSE as default.}

\item{req_vel}{Boolean. Velocity required; when set true, exclude data without a velocity. FALSE as default.}

\item{req_vel_comp}{Boolean. Velocity components required; when set TRUE, exclude data without velocity components. FALSE as default.}

\item{vel_comp}{Boolean. Include velocity components. FALSE as default.}

\item{sort}{String. Sort data on the specified field: “date”, “dist”, “dist-min”, “v-inf”, “v-rel”, “h”, or “object” (default sort order is ascending: prepend “-“ for descending). "date" as default.}

\item{limit}{Integer. Limit data to the first N results (where N is the specified number and must be an integer value greater than zero).}
}
\value{
Fireball atmospheric impact data.
}
\description{
The fireball data API provides a method of requesting specific records from the available data-set. Every successful query will return content representing one or more fireball data records.
See detailed info at: https://ssd-api.jpl.nasa.gov/doc/fireball.html.
}
\examples{
Fireball(date_min = as.Date("2014-01-01"), limit = 20)
Fireball(date_min = as.Date("2015-01-01"), req_alt = TRUE, sort = "-date")
}
