% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_apod_photo.R
\name{get_apod_photo}
\alias{get_apod_photo}
\title{Fetch NASA's Astronomy Picture of the Day}
\usage{
get_apod_photo(api_key, show_image = TRUE)
}
\arguments{
\item{api_key}{Your NASA API key}

\item{show_image}{Logical, whether to display the image in the console. Default is TRUE.}
}
\value{
A list with image URL and description
}
\description{
Fetch NASA's Astronomy Picture of the Day
}
