% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_earth_image.R
\name{get_earth_image}
\alias{get_earth_image}
\title{Get NASA Earth Imagery
Fetches and displays a satellite image for given coordinates and date.}
\usage{
get_earth_image(api_key, lat, lon, date = Sys.Date(), dim = 0.1)
}
\arguments{
\item{api_key}{Your NASA API key.}

\item{lat}{Latitude (-90 to 90).}

\item{lon}{Longitude (-180 to 180).}

\item{date}{Date (YYYY-MM-DD), defaults to today.}

\item{dim}{Image size in degrees, defaults to 0.1.}
}
\value{
List with the saved image file path.
}
\description{
Get NASA Earth Imagery
Fetches and displays a satellite image for given coordinates and date.
}
