% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mars_rover_photos.R
\name{get_mars_rover_photos}
\alias{get_mars_rover_photos}
\title{Get Mars Rover Photos
Fetches image URLs from NASA’s Curiosity Rover for a given sol or Earth date.}
\usage{
get_mars_rover_photos(api_key, sol = NULL, earth_date = NULL)
}
\arguments{
\item{api_key}{NASA API key.}

\item{sol}{Integer for Martian sol (day), optional.}

\item{earth_date}{Date (YYYY-MM-DD) or Date object, optional.}
}
\value{
Vector of image URLs, or NULL if fetch fails.
}
\description{
Get Mars Rover Photos
Fetches image URLs from NASA’s Curiosity Rover for a given sol or Earth date.
}
