% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mars_weather.R
\name{get_mars_weather}
\alias{get_mars_weather}
\title{Get Mars Weather Data
Fetches weather data from NASA’s InSight mission for recent Martian sols.}
\usage{
get_mars_weather(api_key = "DEMO_KEY", version = 1)
}
\arguments{
\item{api_key}{NASA API key, defaults to "DEMO_KEY".}

\item{version}{API version number, defaults to 1.0.}
}
\value{
Data frame with weather details (temperature, pressure, wind) per sol.
}
\description{
Get Mars Weather Data
Fetches weather data from NASA’s InSight mission for recent Martian sols.
}
