% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neo_data.R
\name{get_neo_data}
\alias{get_neo_data}
\title{Get Near-Earth Object Data
Fetches NEO details from NASA’s API for a date range.}
\usage{
get_neo_data(
  start_date,
  end_date,
  api_key = "XFsDPHBjdABhXfquwqnJSfhwEY4rCZI2ev2NDU0K"
)
}
\arguments{
\item{start_date}{Start date (YYYY-MM-DD).}

\item{end_date}{End date (YYYY-MM-DD), max 7 days from start.}

\item{api_key}{NASA API key, defaults to a demo key.}
}
\value{
Data frame with NEO details, or empty if none found.
}
\description{
Get Near-Earth Object Data
Fetches NEO details from NASA’s API for a date range.
}
