% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Insight.R
\name{Insight}
\alias{Insight}
\title{Insight: Mars Weather Service API}
\usage{
Insight(key = Sys.getenv("NASA_TOKEN"), ver = "1.0",
  feedtype = "json", simplified_data_frame = FALSE)
}
\arguments{
\item{key}{String. Your NASA API key, you can enter your key in the function parameter, but it's not recommended. Instead, you'd better save your key in R environment and call it "NASA_TOKEN". Then the function would automatically acquire your key info.}

\item{ver}{String. The version of this API. "1.0" as default.}

\item{feedtype}{String. The format of what is returned. Currently the default is "json" and only "json" works.}

\item{simplified_data_frame}{Boolean. Returns a data frame that only includes date, highest and lowest temperatrue(°F). FALSE as default. One beautiful example that could use this data frame: https://api.nasa.gov/assets/img/general/insight_photo.png.}
}
\value{
List or dataframe of info about weather on Mars.
}
\description{
NASA’s InSight Mars lander takes continuous weather measurements (temperature, wind, pressure) on the
surface of Mars at Elysium Planitia, a flat, smooth plain near Mars’ equator. Summaries of these data
are available at https://mars.nasa.gov/insight/weather/. This API provides per-Sol summary data for each of
the last seven available Sols (Martian Days). As more data from a particular Sol are downlinked from the
spacecraft (sometimes several days later), these values are recalculated, and consequently may
change as more data are received on Earth.
}
\examples{
Insight()
Insight(simplified_data_frame = TRUE)
}
