% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NASA_Image_Video.R
\name{NASA_Image_Video}
\alias{NASA_Image_Video}
\title{NASA_Image_Video: NASA Image and Video Library}
\usage{
NASA_Image_Video(endpoint, nasa_id = NULL, q = NULL, center = NULL,
  description = NULL, description_508 = NULL, keywords = NULL,
  location = NULL, media_type = NULL, page = NULL,
  photographer = NULL, secondary_creator = NULL, title = NULL,
  year_start = NULL, year_end = NULL)
}
\arguments{
\item{endpoint}{String. There are four possible endpoints: "search" (Performing a search), "asset" (Retrieving a media asset’s manifest), "metadata" (Retrieving a media asset’s metadata location), "captions" (Retrieving a video asset’s captions location).}

\item{nasa_id}{String. The media asset's NASA ID. Necessary when endpoint is "asset" or "metadata" or "captions".}

\item{q}{String. Free text search terms to compare to all indexed metadata. Necessary when endpoint is "search".}

\item{center}{String. NASA center which published the media. Optional.}

\item{description}{String. Terms to search for in “Description” fields. Optional.}

\item{description_508}{String. Terms to search for in “508 Description” fields. Optional.}

\item{keywords}{String. Terms to search for in “Keywords” fields. Separate multiple values with commas. Optional.}

\item{location}{String. Terms to search for in “Location” fields. Optional.}

\item{media_type}{String. Media types to restrict the search to. Available types: “image”, “audio”. Separate multiple values with commas. Optional.}

\item{page}{Integer. Page number, starting at 1, of results to get. Optional.}

\item{photographer}{String. The primary photographer’s name. Optional.}

\item{secondary_creator}{String. A secondary photographer/videographer’s name. Optional.}

\item{title}{String. Terms to search for in “Title” fields. Optional.}

\item{year_start}{String. The start year for results. Format: YYYY. Optional.}

\item{year_end}{String. The end year for results. Format: YYYY. Optional.}
}
\value{
Info of images/videos in NASA Image and Video Library.
}
\description{
Use this API(function) to access the NASA Image and Video Library site at http://images.nasa.gov.
}
\details{
See detailed info at https://images.nasa.gov/docs/images.nasa.gov_api_docs.pdf.
}
\examples{
NASA_Image_Video(endpoint = "search", q = "apollo\%2011", description = "moon\%20landing", media_type = "image")
NASA_Image_Video(endpoint = "asset", nasa_id = "as11-40-5874")
NASA_Image_Video(endpoint = "metadata", nasa_id = "as11-40-5874")
NASA_Image_Video(endpoint = "captions", nasa_id = "172_ISS-Slosh")
}
