% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neo_Lookup.R
\name{Neo_Lookup}
\alias{Neo_Lookup}
\title{Asteroids NeoWs: Near Earth Object Web Services}
\usage{
Neo_Lookup(key = Sys.getenv("NASA_TOKEN"), asteroid_id)
}
\arguments{
\item{key}{String. Your NASA API key, you can enter your key in the function parameter, but it's not recommended. Instead, you'd better save your key in R environment and call it "NASA_TOKEN". Then the function would automatically acquire your key info.}

\item{asteroid_id}{Long integer. Asteroid SPK-ID correlates to the NASA JPL small body.}

\item{start_date}{Date. Starting date for asteroid search.}

\item{end_date}{Date. Ending date for asteroid search. 7 days after start_date as default.}

\item{as_dataframe}{Boolean. Return the data in the form of dataframe instead of nested list. False as default.}
}
\value{
Data of Near Earth Object.
}
\description{
NeoWs (Near Earth Object Web Service) is a RESTful web service for near earth Asteroid information.
With NeoWs a user can: search for Asteroids based on their closest approach date to Earth, lookup a specific Asteroid with its NASA JPL small body id, as well as browse the overall dataset.
Dataset: All the data is from the NASA JPL Asteroid team (http://neo.jpl.nasa.gov/).
}
\details{
Neo_Feed: Retrieve a list of Asteroids based on their closest approach date to Earth.

Neo_Lookup: Lookup a specific Asteroid based on its NASA JPL small body (SPK-ID) ID.

Neo_Browse: Browse the overall Asteroid dataset.
}
\examples{
Neo_Feed(start_date = as.Date("2019-11-01"), end_date = as.Date("2019-11-03"), as_dataframe = TRUE)
Neo_Lookup(asteroid_id = 3542519)
Neo_Browse()
}
