% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scout.R
\name{Scout}
\alias{Scout}
\title{Scout: Solar System Dynamics and Center for Near-Earth Object Studies - NEOCP orbits, ephemerides, and impact risk data}
\usage{
Scout(mode = "O", tdes = NULL, plot = NULL, file = NULL,
  orbits = FALSE, n_orbits = 1000, eph_start = "now",
  eph_stop = NULL, eph_step = NULL, obs_code = "500",
  fov_diam = NULL, fov_ra = NULL, fov_dec = NULL, fov_vmag = NULL)
}
\arguments{
\item{mode}{String. Mode to select, "O" or "E". "O" as default.}

\item{tdes}{String. Mode: O, E. Select data for the NEOCP object matching this temporary designation (e.g., "P10uUSw")}

\item{plot}{String. Mode: O. Get plot files for the specified NEOCP object of the selected type ("el"=elements, "ca"=close-approach, "sr"=systematic-ranging). Allowable values: “el”, “ca”, “sr”, “el:ca”, “ca:el:sr” or any similar combination delimited by : where duplicates are ignored.}

\item{file}{String. Mode: O. Get the list of available data files (when file=list) or get the requested data file for the specified NEOCP object; currently, only “mpc” file is available. Allowable values: “list”, “mpc”.}

\item{orbits}{Boolean. Mode: O, E. Get sampled orbits data for the specified NEOCP object. FALSE as default.}

\item{n_orbits}{Number. Mode: O, E. Limit the number of sampled orbits to this value. Range: 1 - 1000. 1000 as default.}

\item{eph_start}{Date/String. Mode: E. Get the ephemeris for the specified NEOCP object at this time UTC. "now" as default.}

\item{eph_stop}{Date. Mode: E. Set the ephemeris stop-time (if specified, requires eph_start and must be later than eph_start).}

\item{eph_step}{String. Mode: E. Set the ephemeris step-size (if specified, requires both eph_start and eph_stop; if not specified, set to the span). Allowable values: "#d", "#h", "#m" (where # is a positive definite integer).}

\item{obs_code}{String. Mode: E. Get the ephemeris for the specified NEOCP object relative to the specified MPC observatory code. "500" as default.}

\item{fov_diam}{Number. Mode: E. Specify the size (diameter) of the field-of-view in arc-minutes. Range: (0, 1800]}

\item{fov_ra}{String. Mode: E. Specify the field-of-view center (R.A. component); requires fov_diam and fov_dec; invalid if eph_stop is set. Allowable values: deg "hh:mm:ss.s" "hh mm ss.s" [0:360) degrees [0:24) hours.}

\item{fov_dec}{String. Mode: E. Specify the field-of-view center (Dec. component): requires fov_diam and fov_ra; invalid if eph-stop is set. Allowable values: +/-deg "dd-mm-ss" "dd mm ss" (absolute value must be less than 90 degrees).}

\item{fov_vmag}{Number. Mode: E. Filter ephemeris results to those with V-magnitude of this value or brighter; requires fov_diam. Range: [0:40].}
}
\value{
Data of NEOCP orbits, ephemerides, and impact risk.
}
\description{
This API provides access to near-realtime results from the CNEOS Scout system.
See detailed info at: https://ssd-api.jpl.nasa.gov/doc/scout.html.
}
\details{
This API supports two query modes:

O - Object Detail Queries

E - Ephemeris Queries
}
\examples{
## Note that the NEOCP object tdes used in the following examples is not guaranteed to still exist in the database.
##### Mode: O
## Scout(tdes = "P10vY9r")
## Scout(tdes = "P10vY9r", plot = "el")
## Scout(tdes = "P10vY9r", orbits = 1)
##### Mode: E
## Scout(mode = "E", tdes = "P10vY9r", eph_start = "2016-09-24T12:00:00")
## Scout(mode = "E", tdes = "P10vY9r", eph_start = "2016-09-25T18:00:00", eph_stop = "2016-09-26T06:00:00", eph_step = "2h")
}
