% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_nasa_images.R
\name{search_nasa_images}
\alias{search_nasa_images}
\title{Search NASA Images
Searches NASA’s image database for a given query.}
\usage{
search_nasa_images(query)
}
\arguments{
\item{query}{Character string to search for.}
}
\value{
List with image metadata (e.g., href, description), or error if failed.
}
\description{
Search NASA Images
Searches NASA’s image database for a given query.
}
