% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sentry.R
\name{Sentry}
\alias{Sentry}
\title{Sentry: Solar System Dynamics and Center for Near-Earth Object Studies - NEO Earth impact risk assessment data}
\usage{
Sentry(mode = "S", spk = NULL, des = NULL, h_max = NULL,
  ps_min = NULL, ip_min = NULL, days = NULL, all = FALSE,
  removed = FALSE)
}
\arguments{
\item{mode}{String. Mode to select, "O" or "S" or "V" or "R". "S" as default.}

\item{spk}{Integer. Mode: O. Select data for the object matching this SPK-ID (e.g., 2029075).}

\item{des}{String. Mode: O. Select data for the object matching this designation (e.g., "29075" or "2000 SG344").}

\item{h_max}{Number. Mode: S, V. Limit data to those with an absolute magnitude, H (weighted-mean for mode-S) less than or equal to this value. Allowable values: [-10:100].}

\item{ps_min}{Integer. Mode: S, V. Limit data to those with a (weighted-mean for mode-S) Palermo scale (PS) greater than or equal to this value. Allowable values: [-20:20].}

\item{ip_min}{Number. Mode: S, V. Limit data to those with a (weighted-mean for mode-S) impact-probability (IP) greater than or equal to this value. Allowable values: [1 - 1e-10].}

\item{days}{Integer. Mode: S, V. Number of days since last observation: limit data to those objects that have been observed within the specified number of days (if the number is negative, limit data to those which have not been observed with the specified number of days). Allowable values: ABS(days) > 6.}

\item{all}{Boolean. Mode: V. Request the complete VI data set. FALSE as default.}

\item{removed}{Boolean. Mode: R. Request the list of removed objects. FALSE as default.}
}
\value{
Data of NEO Earth impact risk assessment.
}
\description{
This API provides access to results from the CNEOS Sentry system.
See detailed info at: https://ssd-api.jpl.nasa.gov/doc/sentry.html.
}
\details{
This API supports four query modes:

O - obtain details related to a specified Sentry object

S - obtain summary data for all available Sentry objects

V - obtain data for all available VIs (Virtual Impactors)

R - obtain summary data for objects removed from Sentry
}
\examples{
##### Mode: O
Sentry(mode = "O", des = "99942")
Sentry(mode = "O", des = "2000\%20SG344")
##### Mode: S
Sentry()
##### Mode: V
Sentry(mode = "V", all = TRUE, ip_min = 1e-3)
##### Mode: R
Sentry(mode = "R", removed = TRUE)
}
