% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Techport.R
\name{Techport}
\alias{Techport}
\title{Techport: API to make NASA technology project data available in a machine-readable format}
\usage{
Techport(key = Sys.getenv("NASA_TOKEN"), id_parameter = NULL,
  all_valid_id = FALSE, update_since = NULL)
}
\arguments{
\item{key}{String. Your NASA API key, you can enter your key in the function parameter, but it's not recommended. Instead, you'd better save your key in R environment and call it "NASA_TOKEN". Then the function would automatically acquire your key info.}

\item{id_parameter}{Integer. The id value of the TechPort record.}

\item{all_valid_id}{Boolean. Print out all valid id of TechPort record if TRUE. FALSE as default.}

\item{update_since}{Date. The start date of returned TechPort record you want.}
}
\value{
Info of technology which NASA is working on.
}
\description{
Techport allows the public to discover the technologies NASA is working on every day to explore space, understand the universe,
and improve aeronautics.
}
\examples{
Techport(id_parameter = 17792)
Techport(all_valid_id = TRUE)
Techport(update_since = as.Date("2016-01-01"))
}
