% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLE.R
\name{TLE}
\alias{TLE}
\title{TLE: Two line element data for earth-orbiting objects at a given point in time.}
\usage{
TLE(satellite_name = NULL, satellite_num = NULL)
}
\arguments{
\item{satellite_name}{String. Performing a search by satellite name.}

\item{satellite_num}{Integer. Retrieving a single TLE record where query is satellite number.}
}
\value{
Data of earth-orbiting objects.
}
\description{
The TLE API provides up to date two line element set records, the data is updated daily from
CelesTrak and served in JSON format. A two-line element set (TLE) is a data format encoding a
list of orbital elements of an Earth-orbiting object for a given point in time.
}
\examples{
TLE(satellite_name = "HTV")
TLE(satellite_num = 43553)
}
